[fix]comment feature fix
